refactor(options): consolidate Usage dashboard helpers
