+++ src/options/widgets/shared/usage/usageChartTypes.ts
@@ -1,9 +1,1 @@
-export interface ChartElements {
-  axis: HTMLElement | null;
-  graph: HTMLElement | null;
-  svg: SVGSVGElement | null;
-  path: SVGPathElement | null;
-  grid: SVGGElement | null;
-  points: SVGGElement | null;
-  xAxis: SVGGElement | null;
-}
+export type { ChartElements } from '@ui/domains/usage-chart';
