+++ src/ui/domains/usage-chart/index.ts
@@ -1,0 +1,20 @@
+export { createUsageChartShell, renderUsageChart } from './usageChart';
+export {
+  DEFAULT_CHART_BOUNDS,
+  buildSmoothPath,
+  computeChartGeometry,
+  formatDateLabel,
+  formatUsageDate,
+  generateTicks,
+  measureChartBounds,
+  parseDateKey,
+  pickLabelIndices,
+  prepareHistory,
+  resolveUsageDateKey,
+  updateAxis,
+  updateGridLines,
+  updatePoints,
+  updateXAxis
+} from './usageChartRenderers';
+export type { ChartElements } from './usageChartTypes';
+export type { ChartGeometry, ChartMeasurements, ChartPoint, TickInfo } from './usageChartRenderers';
