+++ src/ui/domains/usage-chart/usageChartTypes.ts
@@ -1,0 +1,9 @@
+export interface ChartElements {
+  axis: HTMLElement | null;
+  graph: HTMLElement | null;
+  svg: SVGSVGElement | null;
+  path: SVGPathElement | null;
+  grid: SVGGElement | null;
+  points: SVGGElement | null;
+  xAxis: SVGGElement | null;
+}
